% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-package-test-package.R
\docType{package}
\name{r-package-test-package}
\alias{r-package-test}
\alias{r-package-test-package}
\title{r-package-test: R Package Test}
\description{
This package only serves testing purposes.
}
\author{
\strong{Maintainer}: Christian Treffenstaedt \email{treffenstaedt@psych.uni-goettingen.de} (\href{https://orcid.org/0000-0001-5394-0379}{ORCID})

}
\keyword{internal}
